Hide close button on dialogs w. a minimized pos.
When I’m GMing I keep accidentally closing character sheets for
enemies, even though I actually mean to minimize them. This way sheets
that have a minimized position cannot be closed when they’re not
minimized, and I don’t have to jump between the Chat and Journal tabs
as if I was playing a bad version of Frogger.